Update css file to conform to lint rules
Variables are too fancy for old browsers and so are ommitted for now. Instead we'll just do a find/replace on the existing hex values in order to replace duplicated values.

Updated the ordering of values to be alphabetical (which caught a duplicated margin in one rule).

Removed the use of background-position-y, which is apparently too fancy for mainstream use at the moment.