Support multiple RSVP responses
Guests can now respond on behalf of multiple people. If a guest does not
have javascript enabled they can enter responses for a maximum of two
people.

Some of the styling (around the last child of form-labels) was
unnecessary and actually unused, and was harmful when appending
additional guest name inputs.

Additionally guests should ideally not be able to re-submit valid data,
so the submit button is disabled when submitting a response and is not
re-enabled unless something goes wrong.

Fixes #20